collector/requirements.txt
requests
supabase
python-dateutil
flask
flask-cors
python-dotenv
waitress
apscheduler
gunicorn
